refactor: move ProductData and getProductData to utils/products for better organization and type clarity in component imports
