+++ app/lib/utils/products.ts
